Moving univariate_effects plot to plotly and renaming it to slice_plot (adjusting __init__.py)
