+++ src/estimagic/visualization/slice_plot.py
